Backend/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.4.4
python-dotenv==1.0.0
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-MySQLdb==1.0.1
cloudinary==1.32.0
bcrypt==4.0.1